mining-microservice/requirements.txt
boto3>=1.34.0
PyPDF2>=3.0.1
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
lancedb
python-dotenv
langchain